remove vorpal commands
