+++ src/commands/quit.js
