--- src/test/index.js
@@ -1,20 +1,0 @@
-"use strict";
-
-const glob = require("glob");
-const path = require("path");
-
-/**
- * Command option declarations
- * @param {import("commander").Command} program Commander
- * @param {import("moleculer").ServiceBroker} broker Moleculer's Service Broker
- */
-module.exports = function (program, broker) {
-	const files = glob.sync(path.join(__dirname, "*.js"));
-	files.sort();
-	files.forEach((file) => {
-		if (path.basename(file) != "index.js") {
-			const { declaration } = require(file);
-			declaration(program, broker);
-		}
-	});
-};
